Fix op kernel bugs for FPGA tracks close #1004

+++ src/operators/feed_op.h
@@ -53,7 +53,7 @@ class FeedOp : public framework::OperatorBase<DeviceType> {
     auto input_ptr = input->data<float>();
     fpga::format_image(input);
     Tensor *output = param_.Out();
-    auto output_ptr = output->data<half>();
+    auto output_ptr = output->data<float>();
 
     fpga::BypassArgs args = {fpga::DATA_TYPE_FP32};
